termine la conversion de producto_model a producto

+++ src-tauri/migration/src/m20220101_000001_create_table.rs
@@ -73,7 +73,7 @@ impl MigrationTrait for Migration {
                     .col(ColumnDef::new(Producto::Marca).string().not_null())
                     .col(ColumnDef::new(Producto::Variedad).string().not_null())
                     .col(ColumnDef::new(Producto::Presentacion).string().not_null())
-                    .col(ColumnDef::new(Producto::Cantidad).double().not_null())
+                    .col(ColumnDef::new(Producto::Cantidad).float().not_null())
                     .col(ColumnDef::new(Producto::UpdatedAt).date_time().not_null())
                     .to_owned(),
             )
